platform/pyproject.toml
[project]
name = "medai-platform"
version = "1.0.0"
description = "Unified Medical AI Platform — MediScan VLM + Medical LLM + OCR"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Auth
    "bcrypt>=4.2.0",
    "PyJWT>=2.9.0",
    "pyotp>=2.9.0",

    # Security
    "cryptography>=43.0.0",

    # HTTP client (for engine communication)
    "httpx>=0.27.0",

    # Email validation
    "email-validator>=2.2.0",

    # Redis (rate-limiting, caching, sessions)
    "redis[hiredis]>=5.0.0",

    # System monitoring
    "psutil>=5.9.0",

    # Intelligent routing (embedding-based intent classification)
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx",  # for TestClient
    "ruff>=0.7.0",
]
voice = [
    "faster-whisper>=1.1.0",
    "coqui-tts>=0.27.4",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "auth",
    "cache",
    "config",
    "db",
    "gateway",
    "observability",
    "router",
    "search",
    "security",
    "services",
]

[tool.setuptools.package-data]
"*" = ["*.sql", "*.yaml", "*.yml"]
